Move login/logout to subpackage
