+++ cmd/nerdctl/login/login_linux_test.go
@@ -14,7 +14,7 @@
    limitations under the License.
 */
 
-package main
+package login
 
 import (
 	"crypto/rand"
